Add nested graph sample for SizeTransform
Adds a new KDoc sample, `SizeTransformNav`, to demonstrate the usage of `sizeTransform` on a nested navigation graph.

The existing sample has been renamed to `SizeTransformComposable` to better reflect that it demonstrates a `sizeTransform` between two composable destinations. The KDocs for `composable` and `navigation` have been updated to point to the correct samples.

Test: all tests pass
Bug: 452071834
Change-Id: Ifac421044f492059897e47b14343d7b41ac56c94