Add auto-focus support to VerticalStack.
Focus moves are implemented using FocusRequesters to ensure that even
if a focus target is currently underneath the overlap with the item
above it can still gain focus. Similarly, when focus needs to move to
the next item, the current top item may have a focus target in the area
that overlaps with the next item.

Bug: 413429531
Test: VerticalStackTest
Change-Id: I894661875f9dfa31aa7e85791d2de2f45846ea25