Update JetBrains fork version constraints
Update the dependency constraints across multiple modules for the
JetBrains Compose and Lifecycle forks.

This change is necessary to prevent symbol duplication issues that arise
when using older versions of the JetBrains-forked modules. The updated
constraints enforce a minimum version where the JetBrains modules are
published as empty artifacts, depending on the official AndroidX
modules.

Bug: 453194425
Test: manual
Change-Id: I69c1f63f6e546ee10380a8e80cf446da9cc5d457