Update MovableComponent size for main panel.
Reading the main panel size early in the activity lifecycle sometimes returns incorrect values. That leads to attached movable component showing a incorrectly sized move affordance size.

Now, when the layout changes, this listener on main panel's root view updates the `size` of the `mainPanelMovableComponent` to match the size of the `mainPanelEntity`. This ensures the movable area correctly reflects the panel's dimensions after a layout pass.

Test: Manually tested with the test app.
Bug: b/434036574
Change-Id: I4bfd381e755f2cbf7454aeba8817ba73a2623929