Add factory functions for ScatterMap and ScatterSet
Add new factory functions to create a MutableScatterMap from
an existing Map or ScatterMap, and a MutableScatterSet from
an existing Set or ScatterSet.

This provides a convenient and standard way to copy or
convert other collections into new scatter collections.

RelNote: "Add `MutableScatterMap(Map)` and
`MutableScatterSet(Set)` factory functions. Use these
functions to easily create a scatter collection pre-filled
with the contents of an existing collection."

Bug: N/A
Test: ScatterMapTest
Test: ScatterSetTest
Change-Id: I51d70d426606f4deeb2124ee5c51723915d4d3a0