Introduce DisposedException and isDisposed property to Entity
Introduces a new `Entity.DisposedException` that is thrown when client code attempts to use an `Entity` that has already been disposed. This replaces the more generic `IllegalStateException`.

A new `isDisposed` property has also been added to allow for checking if an entity has been disposed.

Relnote: Added `isDisposed` property to Entity. Added new `DisposedException` that is thrown when disposed entity is used, instead of `IllegalStateException`.
Test: Updated existing tests to assert for the new exception and added new tests for the `isDisposed` property.
Change-Id: I983f4f4ef7792c87ec02509602a4a3f47d49fd9a