Fix grandchildren position not being invalidated when there is a layer transformation
Previous logic from onLayoutPositionChanged() was only calling insertOrUpdateTransformedNode(), which is updating this node, but not the subhierarchy. It is crucial to apply the change for the whole subtree. As this operation is pretty expensive, I am trying to optimize it a bit by using cached in rectlist not transformed offsets.

Test: OnGlobalRectChangedTest
Fixes: 450511718
Change-Id: I0aaf2259eddfc7ee2207b6e078a4d561d91d13b2