Add NotificationCompat.ProjectedExtender
This change introduces ProjectedExtender, a new
NotificationComat.Extender for notifications that may be projected to a
Projected device.

The primary feature of this extender is to allow a separate content
intent to be specified for the notification when it is interacted with
on a Projected device. This enables different user experiences on the
host device (e.g., phone) versus the Projected device (e.g., smart
glasses).

The extender provides a setContentIntent() method to set a PendingIntent
that will be fired when the notification is tapped on the Projected
device.

Change-Id: Ie490169cc3941cf142e26a6755b790918e6ee73f