Refactor ShadowDependenciesTask to not use project.provider API
The previous implementation used a `provider {}` block and accessed the values of `excludedModules` and `allowedModules` properties directly. This has been replaced with `flatMap` and `map` to properly chain the providers, ensuring the task configuration remains lazy and efficient.
TESTED: Ran :compose:ui:ui:publish and insspected the ui-inspection.jar and it did not include the repacakged classes. the inspector jar is also back to 1.4 MB.
I also checked the dex content of `androidx.compose.ui.inspection.recompositions.StateReadHadnler.<init>` and line 50 was:
.line 50
    invoke-static {}, Lkotlinx/coroutines/Dispatchers;->getMain()Lkotlinx/coroutines/MainCoroutineDispatcher;
BUG:455970813

Change-Id: If8d1f0bd8bb8595fb9e54ff208f8dbef16c1b684