Annotation changes from lint 32.0.0-alpha10
This upgrade pulls in a UAST update to how field annotations are modeled, with backing fields of properties no longer getting the annotations of the property if they shouldn't apply based on use-site target. For some @PublishedApi properties, that means that backing fields are no longer listed in signature files (which is correct, because unless the property is a const or JvmField, the backing field is not what will be used to access the property).

Test: ./gradlew checkApi
Bug: 450844478
Relnote: N/A
Change-Id: I03baa7871a7c3744c602cd306357580fc64a64f0