Recalculate the rects for the subtree when the layer transformation is removed
When there was a transformation like scale, and it was removed, the whole subtree should be recalculated as all the bounds were affected by the now removed scale. Previous logic was just resetting hasPositionalLayerTransformationsInOffsetFromRoot flag without recalculating the bounds.
It is the bug which made the tests in g3 fail, with it fixed we can re-enable isRectManagerOffsetUsageFromLayoutCoordinatesEnabled.

Bug: 456606764
Bug: 456607982
Test: new tests in OnGlobalRectChangedTest
Change-Id: I8963897a8258f715b3607c7cabf8f53b381b20de