Add trackpad testing APIs
Relnote: "Adds trackpad testing APIs. These are similar to the mouse testing APIs,
    and inject events that mimic the behavior of a tablet or laptop trackpad.
    The primary entry points are SemanticsNodeInteraction.performTrackpadInput and
    MultiModalInjectionScope.trackpad."

Adds trackpad testing APIs to ui-test, which allow verifying the behavior of trackpads.

The primary new API is the TrackpadInjectionScope, which is available through
MultiModalInjectionScope and performTrackpadInput.

The new APIs closely mirror the existing MouseInjectionScope APIs, as trackpads
control the cursor like mice.

Because both the mouse and trackpad control the same cursor, the internals of
InputDispatcher have been adjusted somewhat to allow both types of input to share
state.

On Android, scroll and pinch are notable for being exposed to apps via fake fingers.
The testing APIs mimic these fake finger events, so that they accurately
represent what the platform exposes to apps when performing touchpad gestures.

Fixes: 432326139
Test: new tests

Change-Id: I2ce67fba9c0571ddf3ca97d5994e17062967e61b