Default to unscaled gravity-aligned activity space for Session.
This change updates the default behavior for `Session.create()` to use an unscaled, gravity-aligned activity space.

RelNote: The default value for the `unscaledGravityAlignedActivitySpace` parameter has been changed from `false` to `true` in `Session.create`. This ensures that by default, the activity space is gravity-aligned with a scale of 1 unit = 1 meter.
Bug: b/458171497
Test: Manually with all sample tests apps.
Change-Id: I9f9a2b95fe8e49554a4deeb5bff169bf7db1ad2b