Rename GlanceWearWidget lifecycle methods to onAdded and onRemoved.
Activation/Deactivation events will only be added later on if we have those as part of the widget lifecycle.

Keeping the old methods in the AIDL interface so it doesn't break other libraries currently using those and they will be removed in a future CL.

Bug: 458364657
Test: unit tests
Relnote: Rename GlanceWearWidget lifecycle methods to onAdded and onRemoved.
Change-Id: I62cc39135083554fa5ff711f5bd392e1c98b34d8