Add `rememberRemoteFloat` overload for named values
This commit introduces a new overload for `rememberRemoteFloat` that accepts a `name`, an initial `RemoteFloat` value, and a `domain`.

This allows for the creation of named remote float values, making them overridable in the remote document. The provided initial `RemoteFloat` value is used to set up the expression, which is then assigned a name based on the provided `domain` and `name`.

Test: RemoteStateTest
Bug: 459356166
Change-Id: Idc6acdbce964a1ad2ca74f77bf1b52f4ddf9e0dd