Improve Projected library's test app.
Add elapsed time to DisplayControllerProjectedActivity to force a UI
update.  Without this change, the VISUALS_ON signal that fires right
when the user turns the display on will sometimes fail to show until
the display dims.

Add voice clip playback to DisplayControllerProjectedActivity.
This playback allows verification on audio-only projected devices.

Bug: 459483955
Bug: 459576296
Test: Manually verified test app's CUJs on the device
Change-Id: I29643872b5ab589744267188e4b504b895727319