Guard against reentrant modification when recording derived states in SnapshotStateObserver
SnapshotStateObserver derived state observation is currently structured in a way that makes recursive invocation of apply observers possible. In some cases, it might result in concurrent modification of the scatter map that stores derived state values. This change adds a guard to prevent this type of reentrant read / modifications.

Fixes: 435655844
Bug: 456249373
Bug: 402535073
Test: SnapshotStateObserverTestsCommon
Relnote: Guard against reentrant modification when recording derived states in SnapshotStateObserver
Change-Id: I7b86222f143f4a70dfcc0be9c9bde437ab756e64