Fix the extensions tests failures caused by using different context instances
CameraExtensionCharacteristics has a mechanism to register/unregister a client according to the context used. To ensure that the whole pipeline for the same application context, device id, and attribution tag, the created context is cached and is directly returned for invocations when the input application context, device id, and attribution tag remain the same.

Bug: 459290446, 459343813
Test: ContextUtilTest && VideoCaptureTest in extensionstestapp && manual test
Change-Id: Ib7eca3349abb918aad02381389dfd2c8d549eb10