Refactor ThemeTest to use TestParameterInjector
This change refactors `ThemeTest` by replacing the separate `ThemeComposeTest` and `ThemeViewTest` classes with a single `ThemeTest` class. It uses `TestParameterInjector` to run the same tests for both `Compose` and `View` target players, reducing code duplication.

The `testParameterInjector` dependency is also added to the `build.gradle` file to support this change.

Test: test refactor only
Change-Id: I697520cbbfafed2945d7344b8521e160bf610329