Set `PRESERVE_INDEX` as the default scroll state
This was the only behavior the car host allowed in the past; however,
now apps have the option to turn it off in `SectionedItemTemplate`. As
such, the default should be consistent to the old behavior (as an opt
out) instead of changing the behavior to an opt in.

Bug: 461256656
Fix: 461256656
Test: ./gradlew :car:app:app:test
Relnote: "Change default behavior of upcoming SectionedItemTemplate to
    save scroll state across template refreshes and back actions"
Change-Id: I6c9cc0a686e043604d9097a46747672b837c678c