Refactor: Introduce PdfWriteHandle for document Handle
This change introduces the PdfWriteHandle interface to encapsulate the logic for writing a PdfDocument to a ParcelFileDescriptor.

The EditablePdfDocument now includes a createWriteHandle() method. The SandboxedPdfDocument implementation of this method uses reference counting to manage the lifecycle of the underlying document session.

Each created handle increments the reference count, and closing a handle decrements it. The document and its service connection are only truly closed when the last handle is closed.

Bug: 461370143
Test: build and run on test device
Change-Id: I21a508e5cbbd621290da7480dcc0a01f9e885a1c