Support applying draft edits from fragment
This change introduces methods for applying draft edits from the fragment. This operation internally uses viewModelScope to apply draft edits to the document. Based on the result, either the onApplyEditsSuccess or onApplyEditsFailed callback is invoked. The status of this internal, asynchronous operation can be checked using isApplyEditsInProgress. A handle with write capabilities is provided to the host via the success callback. After the write is complete, the host is responsible for exiting edit mode.

The host can also prompt the user to save or discard unsaved changes using the hasUnsavedChanges flag.

Bug: 461370505
Test: build and run on test device
Change-Id: Ic7b2f4dd632a2feaf0c4fdb1681afa4c446b7dbf