Stop copying over unaffected host test reports
Android Engprod XML test result forwarder is a single thread process.
As we grow the number of modules the forwarding time that runs after
the build continues to grow linearly. To help mitigate this, only copy over
the reports for projects that have been affected by a given change. This
now matches the behavior of androidx_device_tests as well. This will make
reporting time to be proportional to the size of the change.

Test: None
Change-Id: I53a2d27cda9fce6e0c1a3ce1f1ef1547814acf6b