Stop task up to date validation when using -Pandroidx.useMaxDepVersions
-Pandroidx.useMaxDepVersions is used to validate if our libraries and
device tests build and if our host tests pass when the pinned
dependencies are swapped to project dependencies instead.

-Pandroidx.verifyUpToDate is used to ensure that our tasks a configured
correctly so that running them twice does not unexpectely rerun.

-Pandroidx.useMaxDepVersions is always a variation of other
parts of the build, but with a swap of dependencies. The value of testing
to see if the same tasks but with different dependencies rerun is very
low. Spot-checking a few builds this should shorten
androidx_max_deps_versions by about 1-2 minutes (e.g. ab/P109953937)
and androidx_host_tests_max_dep_versions by about 30-45 seconds.

Test: None
Change-Id: I82e8cf2a79e089819765ad811516daa1ad2fcd65