Add Feature for NavigationListener behavior change.
Previously NavigationListener#onNavigationCompleted was not called if
the navigation did not commit. This was changed on the WebView side and
we're adding a Feature so that embedders can tell if they can expect the
new behavior or the old one.

Change-Id: Iddf5e384dd854b4d5b4dd45af0cbf3a03d114c43
Test: Just adding a constant.
Relnote: "Adds NAVIGATION_LISTENER_ON_COMPLETED_FIRES_FOR_NON_COMMITTED
    to WebViewFeature, which can be use to determine whether
    NavigationListener#onNavigationCompleted will be called for
    navigations that don't commit."