Revert "Migrate onPlaced to registerOnLayoutRectChanged()"
This reverts commit cb23213d2487c5b28adb77e621d403f1939a33c1.

Reason for revert: It might be too much of a breaking change for this api to change the timing of the call. The documentation is still claiming it will be called during layout, and the animations sample linked to this api has unintended behavior change after this - the initial animation is different because of that. We might consider deprecating this modifier instead, in order to move our users to a more performant options.

Test: n/a
Relnote: The behavior change for Modifier.onPlaced was reverted and ComposeUiFlags.isNewDispatchingMechanismForOnPlacedEnabled flag was removed.
Change-Id: I78502c61ee2cca02958c7ac3bdda31c7b5f29898