Change import request parameters
Specify credential types and version of the CXF in the import request
class. Previously, these parameters were passed as part of the request
json. However, the library will no longer receive a complete json that
maps to a CXP export request, which contained unnecessary
details which would require additional complexity from the providers.
Take in only the required request parameters that can facilitate the
credential exchange and adhere to CXF. The credential types and the
version will be used to form the requestJson, hidden from the providers,
to be passed to the ProviderEventsApiProvider, as part of the underlying
implementation. The exporters should return credentials and extensions that
 map to the requested credential types and known extensions. They should
only return credentials that they support. If they don't recognize or
support the types in the request, they should ignore credentials for
these types.

Relnote: Refactor ImportCredentialsRequest parameters
Test: Unit tests
Change-Id: I29458b7140bed58de40b47e88a49b6a6117943ca