Add Dp.asRdp() extension function
Adds a `Dp.asRdp()` extension function to convert a `Dp` value directly into a `RemoteDp`. This simplifies the conversion process from `Dp` to `RemoteDp` by avoiding the intermediate step of accessing the `value` property.

A corresponding test is added to verify that the conversion correctly produces the expected pixel value based on the device's density.

Test: RemoteDpTest
Change-Id: I296fa170687f82d383c3307bb5acffcbc11926da