This CL adds registry for mapping stable ids to ephemeral ids.
In case of SandboxedPdfDocument, PdfRenderer holds the data for existing annotations on a document mapped by page level unique ids. On the other hand, view model is responsible for holding document level ids in the format of "AnnotationId_<pageNum>::<unique_id>". This is a guaranteed contract that view model needs to fulfill irrespective of new or existing annotations. Hence we need a way to register existing annotations received from PdfRenderer and map them to the newly formatted ids for retrieval/modification purposes. This change introduces that registry class used to map.

Bug: 462602307
Test: ./gradlew :pdf:pdf-viewer:test
Change-Id: Iae3af2f41c5b100e28440f27b52023a855c28d61