[2Page] Fix selection visibility in TwoPageLayout
This CL fixes an issue where text selection was incorrectly considered invisible in the two-page layout mode of the PDF viewer.

In a two-page layout, a text selection can span across both the left and right pages. In such cases, the bounding box of the entire selection can be misleading. For example, the bottom of the selection on the right page could be vertically higher than the top of the selection on the left page. Therefore, to accurately determine if any part of the selection is visible, the logic is updated to check each individual bounding box of the selection for intersection with the viewport.

This change ensures that the selection is correctly detected as visible, and the action mode is displayed as expected.

Bug: 462363996
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: I50fbd919d63892487e72d1dcc6c4559ed2364d01