Create session on separate thread for MainActivity testapp
The integrated testapp's MainActivity had diskread StrictMode violations
related to the creation of the session on the Main Thread.
Moving the creation to a separate thread avoids this issue.

The Permissions.kt file was refactored into the SessionManager class.

A class was required because the `activity.registerForActivityResult` method has to be called
on the main thread, during (or before) OnCreate(), and then saved as a field.

Bug: 463687170
Test: Tested most test apps with all StrictMode violations turned on
Change-Id: I71835388f90a92946824caa5395d137f3537d755