Migrate importMaven to be run within Gradle process
Starting with Gradle 9.2 ProjectBuilder can no longer be used outside of
Gradle process and we were relying on it in ProjectService of our JVM
application. Instead, move all relevant build logic into a Gradle plugin
so it can run as part of a Gradle process.

Along the way, remove KonanPrebuiltsDownloader that does not actually
work anymore.

Bug: 457413847
Test: ./development/importMaven/importMaven.sh com.android.tools.metalava:metalava:1.0.0-alpha14 --metalava-build-id 14538102
      ./development/importMaven/importMaven.sh import-toml

Change-Id: If1b4158a888b404b91c3151d0c8552945bdbaa80