Remove prebuilt default matcher
Remove prebuilt default matcher from the library. The responsibility of
providing the default logic of matching for exporters will lie on the
optional provider events artifact that the exporters register with. The
optional artifact should check the matcher, and if it is empty, override
it with their own default matcher.

Test: N/A
Change-Id: I2069a15675070032b2a971c164efb767f474563d