Fixing issue with SubspaceDepthSortedSet in SubspaceMeasureAndLayoutDelegate
When a node was being in the process of layout it could have its depth changed due to reparenting when the layout pass was happening and this would change the depth of the nodesPendingLayout. This would cause the IllegalStateException to happen and throw "invalid node depth". By draining the sorted set into a temporary list before executing the layout logic, we make sure to avoid the IllegalStateException.

Test: Unit Tests, added new test to test the specific behavior.
Bug: 465848598
Change-Id: I9db4e29bf2aae645d1c6f61d3c41c3ec77f76d02