Change SpatialRow/Column to accept type-safe alignment.
This is to address API Council feedback and provides the user with type-safety and well-defined behavior for the arguments passed to SpatialRow and SpatialColumn.

Bug: 465188920
Fixes: 465188920
Test: Changed unit tests and test apps
Relnote: The SpatialRow and SpatialColumn APIs that accept a generic SpatialAlignment parameter are deprecated; instead, use the APIs that accept either verticalAlignment or horizontalAlignment for SpatialRow and SpatialColumn, respectively.
Change-Id: Iec390d2f302fe756f7c112ff30e877f74c36589b