Whitebox: create a new activity that combines planes, anchors, and hittest
This new activity combines all of the previous ones and adds camera
rending for all three so that we can add visualizations for planes and
hit testing. Hit testing is now done by tapping the screen on a detected
plane. When successful, an anchor will be visualized on the plane where
the hit was detected.

Bug: 408010870
Test: locally on a Pixel 9

Change-Id: Ibaeacee455998628f4db53fad20677ee0d6ca0c0