Refactor DataStoreImpl versioning for clarity.
This change refactors the Int parameter introduced in aosp/3872400 to be a lambda that can return the correct version based on the Boolean value of the lock status. This change is intended to make the API more readable as each call site to `readDataOrHandleCorruption()` will explicitly be setting the version according to the presence of a lock.

Bug: b/431787506
Test: Should not break existing tests.
Change-Id: Ie701c62f02296c54d371e50f9de814f0c67fd864