Correct MRIR stream configuration on ImageSourceConfig
With multi-output streams (with MultiResolutionImageReader), one has to
create the OutputConfigurations such that each is configured with the
Surface of the internal ImageReader of the MultiResolutionImageReader.
As of right now, the only way to do that is through
OutputConfiguration.createInstancesForMultiResolutionOutput().

This CL does just that, but does still implicitly rely on the guarantee
that the ordering of streams would be maintained.

Bug: 461602496
Test: CameraPipe test app
Change-Id: Icfb1044096da96c920f7e882177dcccfe65cbecc