Pass all source sets to metalava
This updates the source set information created for metalava to include all main source sets. To list the dependencies that are needed for each source set, this finds all compilations which a source set is part of and aggregates the classpath of those compilations because for common source sets it is necessary to include dependencies for all platforms.

Test: Tested running metalava with pending changes to enable multiplatform codebase creation and verified that there were no unresolved types.
Bug: 407737495
Bug: 324624680
Change-Id: Ifb759112f99b7b91e85420ec8f1823df55513de0